+++ CHANGELOG.md
@@ -8,6 +8,22 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 - Initial release
 
+## [0.0.7] - 2022-12-02 18:08:23
+
+### Added
+
+- Modal Button 组件上如果有 结构属性 那么不会格式化
+
+## [0.0.6] - 2022-12-02 16:51:07
+
+### Fixed
+
+- 修复格式化 Panel 组件 title 如果是 jsx 表达式的时候报错问题
+
+### Added
+
+- 增强 Button 组件格式化功能 ，如果之前有 theme 属性并且 不是 `defualt` 就添加 type = 'primary' 属性
+
 ## [0.0.5] - 2022-12-02 12:17:40
 
 ### Fixed